feat-a22-derive-node-confidence-from-edge-evidence
A22 路 B：把节点 confidence 从死常量做成派生量（由边证据决定）。

病灶（实测）：
- 节点 confidence 从不落盘（master_serialize_batch / master_load_state 两处
  序列化路径的节点段都不含该字段）
- 边 confidence 建边时硬编码 0.5f（huarong_topology.c:646）
⇒ edge_conf_to_support(0.5f) ≡ 0.5 ⇒ consistency ≡ 0.5 > 0.3 ⇒ 只升不降
⇒ 线上 3894 样本 min = max = avg = 0.5000
⇒ 感知区三维度知识缺口判据在真实数据上恒 0 命中（结构性死代码）

改法（抄本仓既有口径 template_builder.c:1023，按观测频率分档）：
- 新增 static evidence_to_confidence() / node_edge_evidence()：
  证据 = Σ 出边 weight（唯一活量：实测 0.35–0.99 有真实分布）；
  无出边 ⇒ 最低档 0.05；钳位 [0.05, 0.95]
- knowledge_self_verify() / batch_self_verify() 改为只读 weight 重估 confidence
  （不再读恒为 0.5 的 edges[i].confidence）
- master_load_state() 收尾调 batch_self_verify()，为感知区建立基准
- dialog_system.c 推理收尾：batch_self_verify() 必须排在
  master_consolidate_confidence() 之前，否则反馈加成同轮被覆盖抹掉
- include/perception.h：min_confidence_for_search 由 int 改 float
  （原 PERCEPTION_DEFAULT_CONFIG 给 0.1f 被截断成 0 ⇒ conf < 0 恒假）
- include/multi_topology.h：登记 confidence 现在是派生量、不落盘，禁当持久学习量

验证（armbian-1 aarch64/gcc 13.3.0 + WSL x86_64/gcc 15.2.0，双编译器）：
- make all RC=0 / error 0 / warning 0（WSL 用 make CC=gcc 绕开 ccache 决断读）
- check-version / check-tools / check-wiring 全 PASS（基线 69 条、新增 0）
- make test RC=0，32/32 通过
- 行为：节点置信度 min 0.050 / max 0.900 / avg 0.368；
  维度2 模板缺口 0 → 2128，维度3 拓扑缺口 0 → 1519
- 端到端：perception_tick 首次出词第 100 拍（改前恒返 0）

refs: 待办 A22；工程纪律档案 §7.24

+++ include/perception.h
@@ -50,7 +50,11 @@ typedef enum {
 /** 感觉皮层配置 */
 typedef struct {
     int   max_searches_per_cycle;    /* 每次搜索周期最多搜几个概念 (默认5) */
-    int   min_confidence_for_search;  /* 置信度低于此值才搜 (默认0.3) */
+    /* 🔴 v0.5.40 修 bug：原声明为 `int`，而 PERCEPTION_DEFAULT_CONFIG 给的是
+     * `0.1f` ⇒ **截断为 0** ⇒ `node->confidence < 0` 恒假 ⇒ 拓扑缺口判据
+     * 在结构上永不命中（与 is_valid_query 恒假同类，见工程纪律档案 §7.24）。
+     * 语义上它本来就是「阈值」，必须是 float。 */
+    float min_confidence_for_search;  /* 置信度低于此值才搜 (默认0.1) */
     int   search_timeout_ms;         /* 单次搜索超时(ms) (默认5000) */
     int   cycle_interval_ticks;      /* 搜索周期间隔(脑干tick数) (默认300) */
     int   fallback_interval_ticks;   /* 保底触发间隔 — 无论多忙必搜 (默认1200=20分钟) */
